Implemented temporary fix to ensure Blob storage unit tests run
